Fix: AsymSquidParams constructor iterable args
Some of the args demanded by `AsymSquidParams` constructor could be
single or iterable. For sake of clarity of usage and documenting, and
lack of experience of realization of proper single-iterable union
parameters handling, behaviour was simplified.
Now all parameters that could be iterable are demanded to be passed
as lists.

Signed-off-by: shamil777 <kadyrmetov@phystech.edu>