Fix: `CPW` class `open_end_gap` parameter usage
Parameter was introduced late and took place of the other
crucial input ordered parameter `start` - start point
So, `open_end_gap` (formerly known as `open_end_l`, renamed)
was substituted with value passed as `start` parameter
that has type DPoint
So we has `DPoint` instead of `float` and type errors.
Now it is fixed by shifting `open_end_gap` to be one of the last
parameters to be given.

Signed-off-by: shamil777 <kadyrmetov@phystech.edu>