12QSt: fix simulate_Cqr
Bug was in choosing qubit metal region for port placement.

Qubit metal region is augmented with rectangle for squid. Port was selected on the edge of non-augmented polygon. Hence, during the search for port index there were no augmented polygon's edge that is 10 nm as close to the port. Since the augmented polygon was modified exactly at the port location.
Before this, simulate_Cqr was failing starting with q_idx = [6].